+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # CHANGELOG
 ## Not released
-- Added meta value version to List component [#31](https://github.com/CartoDB/carto-react-lib/pull/31)
+- Add meta value example to List component story [#31](https://github.com/CartoDB/carto-react-lib/pull/31)
 - Mobile fixes and fix loading datasets for free accounts [#38](https://github.com/CartoDB/carto-react-lib/pull/38)
+- Add updateLayer action to cartoSlice [#39](https://github.com/CartoDB/carto-react-lib/pull/39)
 
 ## 1.0.0-beta6 (2020-11-27)
 - Fix CategoryWidgetUI displaying no data while loading [#26](https://github.com/CartoDB/carto-react-lib/pull/26)